Store sessions in Redis
