+++ app/config/config.yml
@@ -58,6 +58,7 @@ framework:
         resource: '%kernel.root_dir%/config/routing.yml'
     secret: '%secret%'
     session:
+        cookie_lifetime: 86400 # 1 day
         name: '%session_name%'
         save_path: '0;660;%kernel.root_dir%/../var/sessions/%kernel.environment%'
     templating:
